Apply build fixes
