--- Tests/FoundationEssentialsTests/Resources/JSON5/spec/strings/no-comma-array.errorSpec
@@ -1,6 +1,0 @@
-{
-    at: 16,
-    lineNumber: 3,
-    columNumber: 5,
-    message: "Expected ']' instead of 'f'"
-}
